Adapt to restructuring changes in Spring Boot 4.0
This commit updates imports and dependencies to adapt to Spring Boot
4.0's restructuring.

Resilence4JMetricsOrderingAutoConfiguration is necessary to correct
the ordering of some Resilience4J auto-configuration that is no
longer correctly ordered by default due to auto-configuration classes
being renamed in Boot 4.0.

This commit also stops using some APIs that have been deprecated
since Boot 3.4 and corrects a test class that was using a mixture of
JUnit 4 and JUnit 5 annotations.

Signed-off-by: Andy Wilkinson <andy.wilkinson@broadcom.com>